LIBS-27 Moved models into a module
Added checkpoints to ignore
